Extracted common parent for HC and M2 GenotypingEngines
